Arrows no hit enemies
Bow and Arrows are now working fully. The Bow can shoot an arrow in which ever direction the player is facing, and the arrows now hit and deal damage to the enemies.

Before Merging this to the Master branch:

Remove input statment in Bow.Update

Remove Bow script from the AnimatedPlayer