Adding art attempt
the assets we found are not all uniform, some of the tiles are 32x32 but other are 31x32. and even some of the correctly sized tlles don't seem to line up correctly.

I also cannot find a way to get ride of the gray boarder around each tile as it's put into the scene... It looks as if it has to do with material the sprite renderer is using, but I can't find a way to make it work as it should..

I've still left all the art in the scene for other to see and review.